+++ Modules/gid.ip/src/test/java/softclub/model/InputPaymentDaoTest.java
@@ -21,10 +21,10 @@ public class InputPaymentDaoTest extends CoreIpModelTester {
 
     @Test
     public void testGetInputSum4Date() throws Exception {
-        Double r = inputPaymentDao.getInputSum4Date(2008, 0, 1);
+        Double r = inputPaymentDao.getInputSum4Date(2008, 0, 1).doubleValue();
         assertNotNull(r);
 
-        r = outputPaymentDao.getMinusSum4Date(1,2008);
+        r = outputPaymentDao.getMinusSum4Date(1,2008).doubleValue();
         assertNotNull(r);
     }
 }
